Move hashify groups code to `groups` module
Stop caching all groups hash
remove `all_groups` method an create hash with no values
  just group names as keys
