+++ src/groups.py
@@ -1,11 +1,31 @@
 
 # all dependancy on nodeattr is to be located in this file
 import config
+from models.config import Config
+from appliance_cli.command_generation import generate_commands
 
 from plumbum import local, ProcessExecutionError
 from tempfile import NamedTemporaryFile
 from click import ClickException
 from re import search
+
+def group_commands(root_command, **kwargs):
+    def __group_commands(callback):
+        groups_hash = __hashify_all_groups(**kwargs)
+        generate_commands(root_command, groups_hash, callback)
+    return __group_commands
+
+# Generates a similar hash to Config hasify funcs - for node groups
+#   {
+#       groupX: **<node>,
+#       ...
+#   }
+def __hashify_all_groups(command = {}):
+    combined_hash = {}
+    for group in list():
+        group_hash = combined_hash.setdefault(group, {})
+        Config._Config__copy_values(command, group_hash, group)
+    return combined_hash
 
 def list():
     groups = __nodeattr(arguments=['-l'])
